requirement.txt
pytorch_metric_learning
opencv

